+++ CHANGELOG.md
@@ -44,7 +44,7 @@ Other guiding principles:
 - **amaru-consensus**: add events and metrics to track the performance of headers processing. ([#1005][])
 - **amaru-ledger**: benchmarks for key volatile db operations (roll forward, switch to fork and context preparation).
 - **amaru-ledger**: add stateful checks on withdrawals (drep delegation requirements + existence of credentials) ([#1011][], [#890][] [#895][])
-- **amaru-consensus**: track pool opcert sequence numbers in the chain store and enforce the Praos rule 
+- **amaru-consensus**: track pool opcert sequence numbers in the chain store and enforce the Praos rule
   that a pool sequence number minus its last known value must be at most 1.
   Sequence numbers are migrated from header already stored in the chain store and otherwise seeded
   from the cardano-node snapshot at bootstrap. ([#1021][])
@@ -265,5 +265,5 @@ Other guiding principles:
 [#1082]: https://github.com/pragma-org/amaru/pull/1082
 [#1095]: https://github.com/pragma-org/amaru/issues/1095
 [#1098]: https://github.com/pragma-org/amaru/pull/1098
-[#1109]: https://github.com/pragma-org/amaru/pull/1109
 [#1101]: https://github.com/pragma-org/amaru/pull/1101
+[#1109]: https://github.com/pragma-org/amaru/pull/1109